empty recycler

+++ app/src/main/java/space/spitsa/beerlist/Beer.kt
@@ -11,15 +11,15 @@ class Beer {
     val description: String? = null
     @SerializedName("image_url")
     val imageUrl: String? = null
-    val abv: Int? = null
-    val ibu: Int? =null
+    val abv: Float? = null
+    val ibu: Float? =null
     @SerializedName("target_fg")
     val targetFg: Int? = null
     @SerializedName("target_og")
-    val targetOg: Int? = null
+    val targetOg: Float? = null
     val ebc: Int? = null
-    val srm: Int? = null
-    val ph: Int? = null
+    val srm: Float? = null
+    val ph: Float? = null
     @SerializedName("attenuation_level")
-    val attenuationLevel: Int? = null
+    val attenuationLevel: Float? = null
 }
